@@ -13,10 +13,12 @@
     }
   ],
   "require": {
-    "php": ">=5.5.0"
+    "php": ">=8.0",
+    "ext-pdo": "*",
+    "ext-oci8": "^3"
   },
   "require-dev": {
-    "phpunit/phpunit": "3.7.*"
+    "phpunit/phpunit": ">=9.5"
   },
   "autoload": {
     "psr-4": {
@@ -24,6 +26,8 @@
     }
   },
   "config": {
-    "github-protocols": ["https"]
+    "github-protocols": [
+      "https"
+    ]
   }
 }
