@@ -1,301 +1,228 @@
 <?php
 
+use Intersvyaz\Pdo\Oci8;
+use Intersvyaz\Pdo\Oci8Exception;
+use PHPUnit\Framework\TestCase;
+
 /**
- * Test class for Pdo_Oci8_Statement.
- * Generated by PHPUnit on 2009-06-02 at 02:18:34.
+ * Тестирование функционала Oracle
  */
-class Pdo_Oci8_StatementTest extends PHPUnit_Framework_TestCase
+class Oci8StatementTest extends TestCase
 {
-    /**
-     * @var    Pdo_Oci8_Statement
-     * @access protected
-     */
-    protected $_object;
+    const DEFAULT_USER = '';
+    const DEFAULT_PWD = '';
+    const DEFAULT_DSN = '';
+
+    /** @var Oci8|null  */
+    protected ?Oci8 $con = null;
 
     /**
-     * Sets up the fixture, for example, opens a network connection.
-     * This method is called before a test is executed.
-     *
-     * @access protected
+     * Настройки перед запуском тестов
      */
-    protected function setUp()
+    public function setUp(): void
     {
-        // $this->_object = new \Intersvyaz\Pdo\Oci8($dsn, array('charset'));
+        $user = getenv('OCI_USER') ?: self::DEFAULT_USER;
+        $pwd = getenv('OCI_PWD') ?: self::DEFAULT_PWD;
+        $dsn = getenv('OCI_DSN') ?: self::DEFAULT_DSN;
+        $this->con = new Oci8($dsn, $user, $pwd, [PDO::ATTR_CASE => PDO::CASE_NATURAL]);
+        $this->con->test = true;
     }
 
     /**
-     * Tears down the fixture, for example, closes a network connection.
-     * This method is called after a test is executed.
-     *
-     * @access protected
+     * Тест проверки, что соединение есть
      */
-    protected function tearDown()
+    public function testObject()
     {
-        unset($this->_object);
+        $this->assertNotNull($this->con);
     }
 
     /**
-     * @todo Implement testExecute().
+     * Удачный постоянный коннект
      */
-    public function testExecute()
+    public function testPersistentConnection()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $user = getenv('OCI_USER') ?: self::DEFAULT_USER;
+        $pwd = getenv('OCI_PWD') ?: self::DEFAULT_PWD;
+        $dsn = getenv('OCI_DSN') ?: self::DEFAULT_DSN;
+        $con = new Oci8($dsn, $user, $pwd, [PDO::ATTR_PERSISTENT => true]);
+
+        $this->assertNotNull($con);
     }
 
     /**
-     * @todo Implement testFetch().
+     * Удачный коннект с параметрами
      */
-    public function testFetch()
+    public function testConnectionWithParameters()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $user = getenv('OCI_USER') ?: self::DEFAULT_USER;
+        $pwd = getenv('OCI_PWD') ?: self::DEFAULT_PWD;
+        $dsn = getenv('OCI_DSN') ?: self::DEFAULT_DSN;
+        $con = new Oci8("$dsn;charset=utf8", $user, $pwd);
+        $this->assertNotNull($con);
     }
 
     /**
-     * @todo Implement testBindParam().
+     * Тест на exception при установлении соединения
      */
-    public function testBindParam()
+    public function testInvalidConnection()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $user = 'pdooci';
+        $pwd = 'pdooci';
+        $str = self::DEFAULT_DSN;
+        $this->expectException(Oci8Exception::class);
+        new Oci8($str, $user, $pwd, [PDO::ATTR_PERSISTENT => true]);
     }
 
     /**
-     * @todo Implement testBindColumn().
+     * Тест на установку и получение атрибута
      */
-    public function testBindColumn()
+    public function testAttributes()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $this->con->setAttribute(PDO::ATTR_AUTOCOMMIT, true);
+        $this->assertTrue($this->con->getAttribute(PDO::ATTR_AUTOCOMMIT));
     }
 
     /**
-     * @todo Implement testBindValue().
-     */
-    public function testBindValue()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testRowCount().
-     */
-    public function testRowCount()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testFetchColumn().
-     */
-    public function testFetchColumn()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testFetchAll().
-     */
-    public function testFetchAll()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testFetchObject().
-     */
-    public function testFetchObject()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testErrorCode().
+     * Тест на получение ошибки, что такой таблицы нет
      */
     public function testErrorCode()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $this->expectException(PDOException::class);
+        $this->expectExceptionCode(942);
+        $this->con->exec("insert into asd (asd) values ('asd')");
     }
 
     /**
-     * @todo Implement testErrorInfo().
+     * Тест на проверку, что драйвер OCI есть
      */
-    public function testErrorInfo()
+    public function testDrivers()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $this->assertTrue(in_array('oci', $this->con->getAvailableDrivers()));
     }
 
     /**
-     * @todo Implement testSetAttribute().
+     * Тест проверки транзакций
      */
-    public function testSetAttribute()
+    public function testInTransaction()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $this->con->beginTransaction();
+        $this->assertTrue($this->con->inTransaction());
+        $this->con->commit();
+        $this->assertFalse($this->con->inTransaction());
     }
 
     /**
-     * @todo Implement testGetAttribute().
+     * Test quotes.
      */
-    public function testGetAttribute()
+    public function testQuote()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $this->assertEquals("'Nice'", $this->con->quote('Nice'));
+        $this->assertEquals("'Not '' string'", $this->con->quote('Not \' string'));
     }
 
     /**
-     * @todo Implement testColumnCount().
+     * Test if returns the last inserted id with a sequence.
      */
-    public function testColumnCount()
+    public function testLastIdWithSequence()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $id = $this->con->lastInsertId('A_PAYS_TRN_SEQ');
+        $this->assertTrue(is_numeric($id));
     }
 
     /**
-     * @todo Implement testGetColumnMeta().
+     *
      */
-    public function testGetColumnMeta()
+    public function testCaseDefaultValue()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $case = $this->con->getAttribute(PDO::ATTR_CASE);
+        $this->assertEquals(PDO::CASE_NATURAL, $case);
     }
 
     /**
-     * @todo Implement testSetFetchMode().
+     * Test setting case.
+     *
+     * @dataProvider caseProvider
      */
-    public function testSetFetchMode()
+    public function testSettingCase()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $case = 'FIELD';
+        $this->con->setAttribute(PDO::ATTR_CASE, $case);
+        $this->assertEquals($case, $this->con->getAttribute(PDO::ATTR_CASE));
+    }
+
+    public function caseProvider(): array
+    {
+        return [
+            [PDO::CASE_LOWER],
+            [PDO::CASE_UPPER],
+        ];
     }
 
     /**
-     * @todo Implement testNextRowset().
+     * Тест создания запроса
      */
-    public function testNextRowset()
+    public function testQuery()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $statement = $this->con->query('SELECT table_name FROM user_tables', null, null);
+        $this->assertInstanceOf(PDOStatement::class, $statement);
     }
 
     /**
-     * @todo Implement testCloseCursor().
+     * Тест установки одного параметра
      */
-    public function testCloseCursor()
+    public function testBindParamSingle()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $stmt = $this->con->prepare('INSERT INTO person (name) VALUES (:a)');
+        $var = 'Joop';
+        $this->assertTrue($stmt->bindParam('a', $var, PDO::PARAM_STR));
     }
 
     /**
-     * @todo Implement testDebugDumpParams().
+     * Тест установки нескольких параметров
      */
-    public function testDebugDumpParams()
+    public function testBindParamMultiple()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $var = 'Joop';
+        $email = 'joop@world.com';
+        $phone = 9639995544;
+        $stmt = $this->con->prepare('INSERT INTO person, email, phone (name) VALUES (:person, :email, :phone)');
+
+        $this->assertTrue($stmt->bindParam(':person', $var, PDO::PARAM_STR));
+        $this->assertTrue($stmt->bindParam(':email', $email, PDO::PARAM_STR));
+        $this->assertTrue($stmt->bindParam(':phone', $phone, PDO::PARAM_INT));
     }
 
-    /**
-     * @todo Implement testCurrent().
-     */
-    public function testCurrent()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
 
     /**
-     * @todo Implement testKey().
+     * Тест установки нескольких значений
      */
-    public function testKey()
+    public function testBindValueMultiple()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $var = 'Joop';
+        $email = 'joop@world.com';
+        $phone = 9639995544;
+        $stmt = $this->con->prepare('INSERT INTO person, email, phone (name) VALUES (:person, :email, :phone)');
+
+        $this->assertTrue($stmt->bindValue(':person', $var, PDO::PARAM_STR));
+        $this->assertTrue($stmt->bindValue(':email', $email, PDO::PARAM_STR));
+        $this->assertTrue($stmt->bindParam(':phone', $phone, PDO::PARAM_INT));
     }
 
-    /**
-     * @todo Implement testNext().
-     */
-    public function testNext()
+    public function testSetConnectionIdentifier()
     {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
+        $expectedIdentifier = 'PDO_OCI8_CON';
 
-    /**
-     * @todo Implement testRewind().
-     */
-    public function testRewind()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
-    }
+        $user = getenv('OCI_USER') ?: self::DEFAULT_USER;
+        $pwd = getenv('OCI_PWD') ?: self::DEFAULT_PWD;
+        $dsn = getenv('OCI_DSN') ?: self::DEFAULT_DSN;
+        $con = new Oci8($dsn, $user, $pwd);
+        $this->assertNotNull($con);
 
-    /**
-     * @todo Implement testValid().
-     */
-    public function testValid()
-    {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-          'This test has not been implemented yet.'
-        );
+        $con->setClientIdentifier($expectedIdentifier);
+        $stmt = $con->query("SELECT SYS_CONTEXT('USERENV','CLIENT_IDENTIFIER') as IDENTIFIER FROM DUAL");
+        $foundClientIdentifier = $stmt->fetchColumn(0);
+        $con->close();
+
+        $this->assertEquals($expectedIdentifier, $foundClientIdentifier);
     }
 }
-?>
